moved routing directly to main.tsx, deleted app.tsx
